requirements.txt
###### Requirements without Version Specifiers ######
streamlit
pandas

sklearn.ensemble